Pulling most recent versions from jrocholl.
In the process of undoing some of my 'local' changes, I lost a bunch of
reversions back to the originals.

This is a commit, verified via

git diff --name-only upstream/master -- *

to verify we've pulled in everything correctly, and that the repository
is a perfect mirror.

Future changes will be organized so that the only differences needed
all live in configuration.scad, rather than being strewn throughout.