Move bearings up to be better positioned to catch the lip of a MK7 drive gear.
When i printed Johann's design, I ended up with the supporting bearings
being poorly aligned with the top edge of the MK7 gear.

To bring it closer to the center of the bearing, I've moved the
bearings nearer the top; this provides better alignment on the MK7's
top lip.